chore: rename android app id
