+++ packages/mobile/android/app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdk 30
     defaultConfig {
-        applicationId 'org.iota.firefly.mobile.alpha'
+        applicationId 'org.iota.firefly.mobile'
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdk 31
         versionCode 1
