+++ packages/mobile/android/app/src/main/java/org/iota/firefly/mobile/MainActivity.java
@@ -1,4 +1,4 @@
-package org.iota.firefly.mobile.alpha;
+package org.iota.firefly.mobile;
 
 import android.os.Bundle;
 import com.getcapacitor.BridgeActivity;
