+++ packages/mobile/android/app/src/main/res/values/strings.xml
@@ -2,6 +2,6 @@
 <resources>
     <string name="app_name">Firefly</string>
     <string name="title_activity_main">Firefly</string>
-    <string name="package_name">org.iota.firefly.mobile.alpha</string>
-    <string name="custom_url_scheme">org.iota.firefly.mobile.alpha</string>
+    <string name="package_name">org.iota.firefly.mobile</string>
+    <string name="custom_url_scheme">org.iota.firefly.mobile</string>
 </resources>
